+++ Sources/EmceeLib/Utils/TestEntriesValidator/TestEntriesValidatorConfiguration.swift
@@ -4,18 +4,18 @@ import Models
 public struct TestEntriesValidatorConfiguration {
     public let simulatorControlTool: SimulatorControlTool?
     public let testDestination: TestDestination
-    public let testEntries: [TestArgFile.Entry]
+    public let testArgFileEntries: [TestArgFile.Entry]
     public let testRunnerTool: TestRunnerTool
 
     public init(
         simulatorControlTool: SimulatorControlTool?,
         testDestination: TestDestination,
-        testEntries: [TestArgFile.Entry],
+        testArgFileEntries: [TestArgFile.Entry],
         testRunnerTool: TestRunnerTool
     ) {
         self.simulatorControlTool = simulatorControlTool
         self.testDestination = testDestination
-        self.testEntries = testEntries
+        self.testArgFileEntries = testArgFileEntries
         self.testRunnerTool = testRunnerTool
     }
 }